requirements.txt
requests~=2.24.0
numpy~=1.18.5
tensorflow-gpu~=2.3.0
Pillow~=7.2.0
sklearn~=0.0
scikit-learn~=0.23.2